requirements.txt
python-telegram-bot==22.*
kubernetes==35.*
